cargo: Update dependencies
Update all dependencies to the latest versions. This required a few
changes, because the API for influx_db_client changed significantly, and
hidapi deprecated a number of functions.

This update was necessary, because hidapi does not support FreeBSD
before 1.1.0, and I figured when updating, why not update to the latest
version.